refactor: centralize decimal digit parsing
Motivation:
PR #4943 review feedback pointed out that Integer and Long kept identical radix-10 digit decoding helpers, and that sign-token variables used inconsistent names across related parse paths.

Modification:
Extract a package-private DecimalParseHelpers.parseDecimalDigit helper for shared ASCII/Unicode decimal digit decoding. Update Integer and Long decimal parse paths to import that helper, and normalize sign-token variables to hasPlusSign/hasNegativeSign while preserving the existing semantic parse flags.

Result:
The duplicated helper implementation is removed from Integer and Long, sign naming is consistent across the affected parse entrypoints, and focused javalib/native parse tests plus cross-version javalib compilation pass.

References:
https://github.com/scala-native/scala-native/pull/4943#discussion_r3537969281
https://github.com/scala-native/scala-native/pull/4943#discussion_r3537986379
https://github.com/scala-native/scala-native/pull/4943#discussion_r3537998999
